Deleted unnecessary files + rename
