--- src/menu/File.js
@@ -1,23 +1,0 @@
-import { app, BrowserWindow } from 'electron';
-
-const fileMenu = {
-  label: 'File',
-  submenu: [
-    {
-      label: 'Reload',
-      accelerator: 'CmdOrCtrl+R',
-      click: () => {
-        BrowserWindow.getFocusedWindow().webContents.reloadIgnoringCache();
-      },
-    },
-    {
-      label: 'Quit',
-      accelerator: 'CmdOrCtrl+Q',
-      click: () => {
-        app.quit();
-      },
-    },
-  ],
-};
-
-export default fileMenu;
