--- src/menu/dev_menu_template.js
@@ -1,30 +1,0 @@
-import { app, BrowserWindow } from 'electron';
-
-const devMenuTemplate = {
-  label: 'Development',
-  submenu: [
-    {
-      label: 'Reload',
-      accelerator: 'CmdOrCtrl+R',
-      click: function() {
-        BrowserWindow.getFocusedWindow().webContents.reloadIgnoringCache();
-      },
-    },
-    {
-      label: 'Toggle DevTools',
-      accelerator: 'Alt+CmdOrCtrl+I',
-      click: function() {
-        BrowserWindow.getFocusedWindow().toggleDevTools();
-      },
-    },
-    {
-      label: 'Quit',
-      accelerator: 'CmdOrCtrl+Q',
-      click: function() {
-        app.quit();
-      },
-    },
-  ],
-};
-
-export default devMenuTemplate;
