+++ tasks/build_app.js
@@ -11,9 +11,8 @@ const destDir = jetpack.cwd('./app');
 
 gulp.task('bundle', () =>
   Promise.all([
-    bundle(srcDir.path('background.js'), destDir.path('background.js')),
-    bundle(srcDir.path('app.js'), destDir.path('app.js')),
-    bundle(srcDir.path('ms_t_preload.js'), destDir.path('ms_t_preload.js')),
+    bundle(srcDir.path('mainApp.js'), destDir.path('mainApp.js')),
+    bundle(srcDir.path('preload.js'), destDir.path('preload.js')),
   ])
 );
